Add MetaPost codeless language module and sample .mp file
This initial commit of MetaPost.plist sets up the structure of a
valid BBEdit CLM file and defines the most basic aspects that make
it MetaPost-specific. (It does NOT include keywords, predefined
constants, and function recognition--those will be added in later
commits.) A sample .mp file is provided just as a convenience for
examining the effectiveness of the CLM file.