Add MetaPost function recognition
For MetaPost we want to recognize 'def/vardef/beginfig' as
"functions" for purposes of navigation and code folding.
As MetaPost is rather unconventional, we simply take the
entire line starting with one of those keywords as "function
name"--so far this seems to work well enough in practice.

I also gratuitously include 'nextfig'--a macro from my own
library that I frequently use (in place of 'beginfig(N)').